requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-jazzmin
gunicorn
psycopg2-binary
whitenoise
python-dotenv
dj-database-url
django-cors-headers 
django-smart-selects